requirements.txt
absl-py==0.13.0
antlr4-python3-runtime==4.8
audioread==2.1.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==5.1.0
filelock==3.4.0
fire==0.4.0
future==0.18.2
fvcore==0.1.3.post20210317
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
huggingface-hub==0.2.1
idna==2.10
importlib-metadata==4.6.1
install==1.3.5
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.3.1
librosa==0.6.0
llvmlite==0.31.0
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numba==0.48.0
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.0
opencv-contrib-python==4.5.3.56
packaging==21.3
pandas==1.1.5
Pillow==8.3.1
pip==21.1.3
portalocker==2.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.21
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.3
PyYAML==5.4.1
quadprog==0.1.8
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
setuptools==59.6.0
six==1.16.0
SoundFile==0.10.3.post1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.61.2
transformers==4.14.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
wheel==0.36.2
yacs==0.1.8
zipp==3.5.0
